requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
